Added full namespaces to routes

+++ routes/web.php
@@ -14,13 +14,13 @@ Route::group(['middleware' => ['web']], function () {
 
         // Resource routes
         Route::group(['middleware' => ['admin_model_exists']], function () {
-            Route::get('/admin/{resource}', 'AdminResourceController@index')->name('admin.resource');
-            Route::get('/admin/{resource}/create', 'AdminResourceController@create')->name('admin.resource.create');
-            Route::post('/admin/{resource}', 'AdminResourceController@store');
-            Route::get('/admin/{resource}/{id}', 'AdminResourceController@show')->name('admin.resource.show');
-            Route::put('/admin/{resource}/{id}', 'AdminResourceController@update')->name('admin.resource.update');
-            Route::patch('/admin/{resource}/{id}', 'AdminResourceController@update');
-            Route::delete('/admin/{resource}/{id}', 'AdminResourceController@destroy')->name('admin.resource.delete');
+            Route::get('/admin/{resource}', 'Matthewbdaly\LaravelAdmin\Http\Controllers\AdminResourceController@index')->name('admin.resource');
+            Route::get('/admin/{resource}/create', 'Matthewbdaly\LaravelAdmin\Http\Controllers\AdminResourceController@create')->name('admin.resource.create');
+            Route::post('/admin/{resource}', 'Matthewbdaly\LaravelAdmin\Http\Controllers\AdminResourceController@store');
+            Route::get('/admin/{resource}/{id}', 'Matthewbdaly\LaravelAdmin\Http\Controllers\AdminResourceController@show')->name('admin.resource.show');
+            Route::put('/admin/{resource}/{id}', 'Matthewbdaly\LaravelAdmin\Http\Controllers\AdminResourceController@update')->name('admin.resource.update');
+            Route::patch('/admin/{resource}/{id}', 'Matthewbdaly\LaravelAdmin\Http\Controllers\AdminResourceController@update');
+            Route::delete('/admin/{resource}/{id}', 'Matthewbdaly\LaravelAdmin\Http\Controllers\AdminResourceController@destroy')->name('admin.resource.delete');
         });
     });
 });
